+++ src/typedal/helpers.py
@@ -74,7 +74,7 @@ def all_dict(cls: type) -> AnyDict:
     return dict(ChainMap(*(c.__dict__ for c in reversed_mro(cls))))  # type: ignore
 
 
-def all_annotations(cls: type, _except: typing.Iterable[str] = None) -> dict[str, type]:
+def all_annotations(cls: type, _except: typing.Optional[typing.Iterable[str]] = None) -> dict[str, type]:
     """
     Wrapper around `_all_annotations` that filters away any keys in _except.
 
@@ -322,3 +322,31 @@ def get_field(field: "TypedField[typing.Any] | Field") -> "Field":
         "Field",
         field,  # Table.field already is a Field, but cast to make sure the editor knows this too.
     )
+
+
+class classproperty:
+    """
+    Combination of @classmethod and @property.
+    """
+
+    def __init__(self, fget: typing.Callable[..., typing.Any]) -> None:
+        """
+        Initialize the classproperty.
+
+        Args:
+            fget: A function that takes the class as an argument and returns a value.
+        """
+        self.fget = fget
+
+    def __get__(self, obj: typing.Any, owner: typing.Type[T]) -> typing.Any:
+        """
+        Retrieve the property value.
+
+        Args:
+            obj: The instance of the class (unused).
+            owner: The class that owns the property.
+
+        Returns:
+            The value returned by the function.
+        """
+        return self.fget(owner)