+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 <!--next-version-placeholder-->
 
+## v3.14.0 (2025-05-15)
+
+### Feature
+
+* `db.find_model` to get the registered TypedTable class for a specific 'table_name' ([`c645303`](https://github.com/trialandsuccess/TypeDAL/commit/c645303a85c96735f48eafaaeb20f867b977686f))
+
+## v3.13.1 (2025-04-28)
+
+### Fix
+
+* Pass select kwargs via `.column()` - so you can do e.g. `distinct=True` ([`e5bc168`](https://github.com/trialandsuccess/TypeDAL/commit/e5bc168b90d6a5d214f049de0ef31e544214cc23))
+
+## v3.13.0 (2025-04-28)
+
+### Feature
+
+* Adding `_once` hooks ([`a69fbb3`](https://github.com/trialandsuccess/TypeDAL/commit/a69fbb361cdfec9352fca503206299c5bbb940d2))
+
 ## v3.12.2 (2025-04-25)
 
 ### Fix